Refactor payment handling and enhance exception messages
