+++ app/model/Payment/Exception/PaymentHasNoEmails.php
@@ -12,6 +12,6 @@ class PaymentHasNoEmails extends Exception
 {
     public static function withName(string $name): self
     {
-        return new self(sprintf('Payment "%s" does not have any e-mail address filled', $name));
+        return new self(sprintf('Platba "%s" nemá vyplněný žádný e-mail', $name));
     }
 }
